+++ src/libs/bputils/src/Utils/api/bpfile.h
@@ -333,7 +333,7 @@ namespace bp {
          *  \param      path [IN] - sourc path
          *  \returns    file size
          */
-        size_t size(const Path& path);
+        uintmax_t size(const Path& path);
 
         /** Whether the item specified by the path exists.
          *  Uses boost::filesystem::exists().
